+++ builtin/bins/provider-dme/main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"github.com/hashicorp/terraform/builtin/providers/dnsmadeeasy"
+	"github.com/hashicorp/terraform/builtin/providers/dme"
 	"github.com/hashicorp/terraform/plugin"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
-		ProviderFunc: dnsmadeeasy.Provider,
+		ProviderFunc: dme.Provider,
 	})
 }
