+++ builtin/providers/dme/provider.go
@@ -1,4 +1,4 @@
-package dnsmadeeasy
+package dme
 
 import (
 	"os"
