+++ website/source/assets/stylesheets/_docs.scss
@@ -9,6 +9,7 @@ body.page-sub{
 body.layout-atlas,
 body.layout-consul,
 body.layout-dnsimple,
+body.layout-dme,
 body.layout-cloudflare,
 body.layout-cloudstack,
 body.layout-google,
