+++ website/source/docs/providers/dme/index.html.markdown
@@ -1,7 +1,7 @@
 ---
-layout: "dnsmadeeasy"
+layout: "dme"
 page_title: "Provider: DNSMadeEasy"
-sidebar_current: "docs-dnsmadeeasy-index"
+sidebar_current: "docs-dme-index"
 description: |-
   The DNSMadeEasy provider is used to interact with the resources supported by DNSMadeEasy. The provider needs to be configured with the proper credentials before it can be used.
 ---
