+++ website/source/layouts/dme.erb
@@ -1,0 +1,26 @@
+<% wrap_layout :inner do %>
+    <% content_for :sidebar do %>
+        <div class="docs-sidebar hidden-print affix-top" role="complementary">
+            <ul class="nav docs-sidenav">
+                <li<%= sidebar_current("docs-home") %>>
+                <a href="/docs/index.html">&laquo; Documentation Home</a>
+                </li>
+
+                <li<%= sidebar_current("docs-dme-index") %>>
+                <a href="/docs/providers/dme/index.html">DNSMadeEasy Provider</a>
+                </li>
+
+                <li<%= sidebar_current("docs-dme-resource") %>>
+                <a href="#">Resources</a>
+                <ul class="nav nav-visible">
+                    <li<%= sidebar_current("docs-dme-resource-record") %>>
+                    <a href="/docs/providers/dme/r/record.html">dme_record</a>
+                    </li>
+                </ul>
+                </li>
+            </ul>
+        </div>
+    <% end %>
+
+    <%= yield %>
+    <% end %>
