+++ website/source/layouts/docs.erb
@@ -128,6 +128,10 @@
 					<a href="/docs/providers/do/index.html">DigitalOcean</a>
                     </li>
 
+					<li<%= sidebar_current("docs-providers-dme") %>>
+					<a href="/docs/providers/dme/index.html">DNSMadeEasy</a>
+					</li>
+
 					<li<%= sidebar_current("docs-providers-dnsimple") %>>
 					<a href="/docs/providers/dnsimple/index.html">DNSimple</a>
 					</li>
